# Add Cython implementation for VOTable binary converters