--- a/astropy/io/votable/setup_package.py
+++ b/astropy/io/votable/setup_package.py
@@ -1,7 +1,9 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
+
 from os.path import join
 
+import numpy as np
 from setuptools import Extension
 
 
@@ -13,5 +15,10 @@ def get_extensions(build_type="release"):
             "astropy.io.votable.tablewriter",
             [join(VO_DIR, "tablewriter.c")],
             include_dirs=[VO_DIR],
-        )
+        ),
+        Extension(
+            "astropy.io.votable.fast_converters",
+            [join(VO_DIR, "fast_converters.pyx")],
+            include_dirs=[np.get_include(), VO_DIR],
+        ),
     ]