--- a/docs/changes/io.votable/18454.perf.rst
+++ b/docs/changes/io.votable/18454.perf.rst
@@ -1,0 +1,1 @@
+Improve performance of Binary parsing by moving converters to Cython