@pllim 2025-07-23
Hmm looks like @eerovaher proposed an alternative at https://github.com/astropy/astropy/pull/18455 . Are you interested to have a look, @stvoutsin ? Thanks, all!