@pllim 2025-07-24
https://github.com/astropy/astropy-benchmarks/pull/141 is merged so theoretically next push would trigger a new benchmark run with the new relevant benchmarks. 🤞