@pllim 2025-07-26
Benchmark run looking good. Thanks!

| Change   | Before [db7a3e2e]    | After [30bfc4a5]    |   Ratio | Benchmark (Parameter)                                     |
|----------|----------------------|---------------------|---------|-----------------------------------------------------------|
| -        | 2.82±0s              | 2.00±0s             |    0.71 | votable.TimeVOTableShortStrings.time_short_strings_binary |
| -        | 3.09±0s              | 2.03±0.01s          |    0.66 | votable.TimeVOTableLongStrings.time_long_strings_binary2  |
| -        | 13.6±0.09ms          | 8.81±0.04ms         |    0.65 | votable.TimeVOTableSmallOverhead.time_small_binary2       |
| -        | 10.4±0.1ms           | 6.27±0.03ms         |    0.61 | votable.TimeVOTableSmallOverhead.time_small_binary        |
| -        | 5.60±0s              | 3.35±0.01s          |    0.6  | votable.TimeVOTableMixed.time_mixed_binary2               |
| -        | 4.43±0.01s           | 2.26±0s             |    0.51 | votable.TimeVOTableNumeric.time_numeric_binary2           |
| -        | 3.51±0.01s           | 1.69±0s             |    0.48 | votable.TimeVOTableNumeric.time_numeric_binary            |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.